requirements_test.txt
# Strictly for tests
pytest==8.3.4
pytest-homeassistant-custom-component==0.13.201
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1

# From our manifest.json for our custom component
aiohttp>=3.8.0
